:recycle: refactor: move components to src/components and fix all imports
