+++ src/components/capabilities/lang-capabilities.module.css
