+++ src/components/faq-component/faq-component.module.css
