+++ src/components/faq-component/faq-component.tsx
