+++ src/components/slynxHeroSection/SlynxLandingHeroSection.tsx
