+++ src/components/slynxHeroSection/WhySlynxContainer.tsx
