+++ src/components/slynxHeroSection/slynxHero.module.css
